Corected bug on GPIO pin. When setted to 0 were causing bad I2C behaviour. Some I2C debugging info added. (this is temporary)

+++ Acopter32-STM32F4/hardware/STM32F4xx_DSP_StdPeriph_Lib_V1.1.0/Libraries/CMSIS/Include/arm_math.h
@@ -5886,7 +5886,7 @@ extern "C"
       /* Iniatilize output for below specified range as least output value of table */
       y = pYData[0];
     }
-    else if(i >= S->nValues)
+	  else if((unsigned)i >= S->nValues)
     {
       /* Iniatilize output for above specified range as last output value of table */
       y = pYData[S->nValues - 1];
